Feat: 결제 후 검토 기능

+++ src/main/java/LikeLion/UnderTheCBackend/repository/ShoppingListRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ShoppingListRepository extends JpaRepository<ShoppingList, Integer> {
-    List<ShoppingList> findByBuyerId(Integer buyer_id);
+public interface ShoppingListRepository extends JpaRepository<ShoppingList, Long> {
+    List<ShoppingList> findByBuyerId(Long buyer_id);
 }
